pyproject.toml
[project]
name = "pbl-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "firecrawl-py>=4.14.0",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]
